Extracted tileset class
